BE_221110_debug : UnityAlbumResponseDTO position -> id로 변환하도록 변경 (완료)

+++ backend/src/main/java/com/ssafy/zip/dto/response/UnityAlbumResponseDTO.java
@@ -1,4 +1,4 @@
 package com.ssafy.zip.dto.response;
 
-public record UnityAlbumResponseDTO(Long id, Long albumId, Long pictureId) {
+public record UnityAlbumResponseDTO(Integer id, Long albumId, Long pictureId) {
 }
